# PERF: Performance improvement value_counts for masked arrays