--- a/doc/source/whatsnew/v1.6.0.rst
+++ b/doc/source/whatsnew/v1.6.0.rst
@@ -100,8 +100,8 @@ Deprecations
 
 Performance improvements
 ~~~~~~~~~~~~~~~~~~~~~~~~
--
--
+- Performance improvement for :meth:`Series.value_counts` with nullable dtype (:issue:`40000`)
+- Performance improvement for :class:`Series` constructor passing integer numpy array with nullable dtype (:issue:`40000`)
 
 .. ---------------------------------------------------------------------------
 .. _whatsnew_160.bug_fixes: