--- a/pandas/core/arrays/masked.py
+++ b/pandas/core/arrays/masked.py
@@ -948,31 +948,22 @@ class BaseMaskedArray(OpsMixin, ExtensionArray):
         )
         from pandas.arrays import IntegerArray
 
+        keys, value_counts = algos.value_counts_arraylike(
+            self._data, dropna=True, mask=self._mask
+        )
+
         if dropna:
-            keys, counts = algos.value_counts_arraylike(
-                self._data, dropna=True, mask=self._mask
-            )
-            res = Series(counts, index=keys)
+            res = Series(value_counts, index=keys)
             res.index = res.index.astype(self.dtype)
             res = res.astype("Int64")
             return res
 
-        # compute counts on the data with no nans
-        data = self._data[~self._mask]
-        value_counts = Index(data).value_counts()
-
-        index = value_counts.index
-
         # if we want nans, count the mask
-        if dropna:
-            counts = value_counts._values
-        else:
-            counts = np.empty(len(value_counts) + 1, dtype="int64")
-            counts[:-1] = value_counts
-            counts[-1] = self._mask.sum()
+        counts = np.empty(len(value_counts) + 1, dtype="int64")
+        counts[:-1] = value_counts
+        counts[-1] = self._mask.sum()
 
-            index = index.insert(len(index), self.dtype.na_value)
-
+        index = Index(keys, dtype=self.dtype).insert(len(keys), self.dtype.na_value)
         index = index.astype(self.dtype)
 
         mask = np.zeros(len(counts), dtype="bool")